Refactor TestBases concerning flash tests
- Move `FormTemplateTestMixin` to tests.base
- Move `AppInitialized` to `test.base`
- Fix imports accordingly
- Use new flash asserts in hss integration tests
